run tests in node without spinning up a browserwindow
Test runs using a browserwindow are significantly slower than running under node (at least 2x). 

These changes allow punting to jest's node environment when set with the docblock tag `@jest-environment node`. 

This is similar to disabling jsdom outside an electron environment.
